Slight Changes + Unaccessible Section
You currently won't be able to access the user portal due to development on the Sign In / Create Account sections. (You may hit a dead end)

When you hit an error on the price plan section, you can now see the distance of your selected custom journey and the maximum distance of the aircraft. This should make it much easier to find out why.